Implemented setIV()
setIV will ask for an Initilization Vector. If you don't
provide one, it will generate one and store it into the
cipherText.

On decryption, if an IV is not provided you need to manually
get the IV from the first bytes of the cipherText and then
strip those bytes from the ciphertext